backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.2.0
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0



